% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_lb.R
\name{boxplot_lb}
\alias{boxplot_lb}
\title{Box plot with the data points overlaid}
\usage{
boxplot_lb(
  data,
  group,
  value,
  highlight_group = NULL,
  add_mean = FALSE,
  use_palette = FALSE
)
}
\arguments{
\item{data}{A data frame containing the groups and corresponding values}

\item{group}{A vector containing the grouping variables}

\item{value}{The corresponding values for each group}

\item{highlight_group}{Logical. TRUE to emphasize a group in the plot by alpha}

\item{add_mean}{Add the mean value as a point to the plot}

\item{use_palette}{Logical. TRUE to use colorblind-safe Okabe–Ito color palette}
}
\value{
A ggplot2 object
}
\description{
Box plots can sometimes mask the distribution of data. Create a boxplot with the data points overlaid. Note that the outliers are turned off since the data points are overlaid.
}
\examples{
data <- data.frame(group = rep(c("a", "b", "c")),
                   value = c(rnorm(30, mean = 6, sd = 5),
                             rnorm(30, mean = 2, sd = 4),
                             rnorm(30, mean = 12, sd = 7)))

boxplot_lb(data = data, group = group, value = value,
           use_palette = TRUE, add_mean = TRUE, highlight_group = "b")


}
