% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_ellipse.R
\name{confidence_ellipse}
\alias{confidence_ellipse}
\title{Confidence ellipse for multivariate multiple regression models}
\usage{
confidence_ellipse(mod, newdata, level = 0.95, ggplot = TRUE)
}
\arguments{
\item{mod}{The fitted \code{lm} model object}

\item{newdata}{A required data frame in which to look for variables with which to predict.}

\item{level}{The confidence level. Default is 0.95}

\item{ggplot}{logical indicating if plot should be created using ggplot2. Default is TRUE. Other plot is created with base graphics.}
}
\value{
A plot with one response on the x-axis and the other on the y-axis, with a single point representing the predicted mean values for both responses and an ellipse indicating the uncertainty in the prediction.
}
\description{
Function that plots a confidence ellipse for two responses in a  multivariate multiple regression model fitted with \code{lm}. The ellipse is created using the \code{ellipse} function from the car package.
}
\examples{
m <- lm(cbind(TOT, AMI) ~ GEN + AMT, data = ami)
nd <- data.frame(GEN = 1, AMT = 1200)
confidence_ellipse(mod = m, newdata = nd)

}
\references{
Johnson R and Wichern D (2007). Applied Multivariate Statistical Analysis, Sixth Edition. Prentice-Hall. (page 399)

Fox J, Weisberg S (2019). \emph{An R Companion to Applied Regression}, Third edition. Sage, Thousand Oaks CA. \url{https://www.john-fox.ca/Companion/}.
}
