% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_chart.R
\name{dot_chart}
\alias{dot_chart}
\title{Dot Chart}
\usage{
dot_chart(data, formula = NULL, segments = TRUE)
}
\arguments{
\item{data}{a data frame containing the classifying variables and the corresponding counts for each combination of the classifying variables. This kind of data frame is easily created using \code{as.data.frame.table()} on a table object created with \code{xtabs()}. (See examples below.) The minimum number of columns is two. The maximum number of columns is four.}

\item{formula}{a formula with the left-hand side containing the counts and the right-hand side containing the cross-classifying variables. The order of the classifying variables determines their role in the plot. The first variable will be on the y-axis. The second variable will form the rows in a call to \code{facet_grid()}. If the second variable is also the last variable, it will be the faceting variable in a call to \code{facet_wrap}. The third variable will form the columns in a call to \code{facet_grid()}. If no formula is provided, the function assumes the data frame was created with \code{as.data.frame.table()} and looks for counts in the last column.}

\item{segments}{Add segments from axis to dots. Default is TRUE.}
}
\value{
Returns a Cleveland dot chart.
}
\description{
Create a Cleveland dot plot using ggplot2. This is useful for visualizing categorical data.
}
\examples{
# 3 dimensions
dot_chart(Freq ~ Dept + Gender + Admit, data = UCBAdmissions)

# 2 dimensions
xtabs(Freq ~ Dept + Gender, data = UCBAdmissions) |>
  as.data.frame() |>
  dot_chart()
xtabs(Freq ~ Dept, data = UCBAdmissions) |>
  as.data.frame() |>
  dot_chart()

# example with proportions
proportions(UCBAdmissions, margin = c(2,3)) |>
  as.data.frame(responseName = "Proportion") |>
  dot_chart(formula = Proportion ~ Dept + Gender + Admit)

# turn off segments
proportions(UCBAdmissions, margin = c(2,3)) |>
  as.data.frame(responseName = "Proportion") |>
  dot_chart(formula = Proportion ~ Dept + Gender + Admit,
            segments = FALSE)
}
\references{
\url{https://hbiostat.org/bbr/descript#categorical-variables-1}
}
\seealso{
\code{\link[=dotchart]{dotchart()}} for base R version.
}
