% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_lb.R
\name{theme_lb}
\alias{theme_lb}
\title{Lauren Brideau's ggplot2 theme}
\usage{
theme_lb(
  title_face = "bold",
  title_size = 12,
  axis_color = "black",
  background_color = "white",
  paper = "white",
  ink = "black",
  base_family = "sans",
  base_size = 12,
  text_size = 10,
  base_line_size = 0.5,
  base_rect_size = 0.5,
  use_palette = FALSE
)
}
\arguments{
\item{title_face}{Character. Font for plot title}

\item{title_size}{Numeric. Font size of title}

\item{axis_color}{Character. Color for axis text and lines}

\item{background_color}{Character. Fill color for the plot background}

\item{paper}{Character. Fill color for panel background}

\item{ink}{Character. Color for all text elements}

\item{base_family}{Character. Font family for all text elements}

\item{base_size}{Numeric. Base font size for text elements}

\item{text_size}{Numeric. Font size for axis text, legend text, and caption}

\item{base_line_size}{Numeric. Base line width for plot lines}

\item{base_rect_size}{Numeric. Border width for rectangular elements}

\item{use_palette}{Logical. TRUE to use color lb_palette. Use with 7 groups
or less. Default is FALSE. The lb_palette consists of the following colors:
\code{c("blue4", "darkorange", "goldenrod1", "gray62", "peachpuff", "peru", "slategray2")}}
}
\value{
A ggplot2 theme that can be added to a ggplot2 object
}
\description{
This function applies a theme to ggplot2 plots. The theme is based on the preferences of Lauren Brideau, a researcher at the University of Virginia.
}
\examples{
library(ggplot2)
p1 <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p1 + theme_lb()

p2 <- ggplot(mtcars, aes(mpg, wt, color = factor(am))) + geom_point()
p2 + labs(color = "AM") + theme_lb(use_palette = TRUE)
}
\seealso{
\itemize{
\item \href{https://ggplot2.tidyverse.org/}{ggplot2 website} - Official ggplot2 documentation
}
}
