% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nmds.R
\name{plot_nmds}
\alias{plot_nmds}
\title{Plot an NMDS object with ggplot2}
\usage{
plot_nmds(nmds, group = NULL)
}
\arguments{
\item{group}{A vector indicating group membership for each observation}

\item{NMDS}{A \code{metaMDS} or \code{monoMDS} object from the vegan package}
}
\value{
A ggplot2 object
}
\description{
Create a custom NMDS plot using ggplot2
}
\examples{
library(ggplot2)
library(vegan)

comm.df <- data.frame(group = rep(c("control", "treatment"), each = 10),
                     matrix(rpois(20*10, lambda = 3), nrow = 20, ncol = 10))

nmds_result <- vegan::metaMDS(comm.df[,2:11], k = 2, trymax = 50)

plot_nmds(nmds_result, group = comm.df$group)

}
\seealso{
\itemize{
\item \href{https://ggplot2.tidyverse.org/}{ggplot2 website} - Official ggplot2 documentation
\item \href{https://cran.r-project.org/web/packages/vegan/vegan.pdf}{vegan reference manual}
}
}
