% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdot_plot.R
\name{vdot_plot}
\alias{vdot_plot}
\title{Dot plot and violin plot}
\usage{
vdot_plot(formula, data)
}
\arguments{
\item{formula}{a model formula of the form y ~ x or, if plotting by groups, y ~ x | z, where z evaluates to a factor or other variable dividing the data into groups. If x or z is not a factor, the function converts them to a factor.}

\item{data}{data frame within which to evaluate the formula.}
}
\value{
Returns a Wilkinson dot chart overlaying a violin plot.
}
\description{
Create a Wilkinson dot plot overlaying a violin plot using ggplot2. The Wilkinson dot plot shows each observation as a dot. The violin plot provides a distribution density estimate. This function basically implements \href{https://hbiostat.org/bbr/descript#fig-descript-vplot}{Fig 4.26} from Frank Harrell's \href{https://hbiostat.org/bbr/}{Biostatistics for Biomedical Research book}.
}
\examples{
vdot_plot(mpg ~ am, mtcars)
vdot_plot(mpg ~ am | vs, mtcars)
}
